requirements.txt
# Core dependencies
pandas==1.5.3
numpy==1.24.3
scikit-learn==1.3.0
psycopg2-binary==2.9.6
faker==18.9.0

# API and web
fastapi==0.100.0
uvicorn==0.23.1
pydantic==2.0.2

# Database and ORM
sqlalchemy==2.0.18
alembic==1.11.1

# Data visualization
matplotlib==3.7.1
seaborn==0.12.2
plotly==5.15.0

# Testing
pytest==7.4.0
pytest-cov==4.1.0

# Development tools
black==23.7.0
flake8==6.0.0
pre-commit==3.3.3

# Notebooks
jupyter==1.0.0
ipykernel==6.24.0

# Environment management
python-dotenv==1.0.0

# Documentation
sphinx==7.0.1
sphinx-rtd-theme==1.3.0

